Move common functions to util function
